Move the data package from spi to api
