+++ core/src/it/java/org/seedstack/seed/core/fixtures/TestDataImporter3.java
@@ -7,8 +7,8 @@
  */
 package org.seedstack.seed.core.fixtures;
 
-import org.seedstack.seed.spi.data.DataImporter;
-import org.seedstack.seed.spi.data.DataSet;
+import org.seedstack.seed.DataImporter;
+import org.seedstack.seed.DataSet;
 
 import java.util.ArrayList;
 import java.util.List;
