+++ core/src/main/java/org/seedstack/seed/core/internal/data/DataManagerImpl.java
@@ -14,8 +14,8 @@ import com.google.common.collect.Lists;
 import com.google.inject.Injector;
 import org.seedstack.seed.DataManager;
 import org.seedstack.seed.SeedException;
-import org.seedstack.seed.spi.data.DataExporter;
-import org.seedstack.seed.spi.data.DataImporter;
+import org.seedstack.seed.DataExporter;
+import org.seedstack.seed.DataImporter;
 
 import javax.inject.Inject;
 import java.io.*;
