+++ core/src/test/java/org/seedstack/seed/core/fixtures/TestDataExporter2.java
@@ -8,7 +8,7 @@
 package org.seedstack.seed.core.fixtures;
 
 import com.google.common.collect.Lists;
-import org.seedstack.seed.spi.data.DataExporter;
+import org.seedstack.seed.DataExporter;
 
 import java.util.Iterator;
